requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
cloudinary==1.44.1
dj-database-url==3.0.1
Django==5.2.5
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-sendgrid-v5==1.3.0
djangorestframework==3.16.1
ecdsa==0.19.1
gunicorn==23.0.0
idna==3.10
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
python-http-client==3.3.7
requests==2.32.5
sendgrid==6.12.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
whitenoise==6.9.0
